requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tortoise-orm==0.20.0
aerich==0.7.2
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiosmtplib==3.0.1
deep-translator==1.11.4
aiofiles==23.2.1
pillow==10.1.0
requests==2.31.0